requirements.txt
psycopg
django
django_rest_framework
djangorestframework
django-rest-swagger
celery
redis